Fixed wrong sentence in README.md
I was confused with the part of the sentence below the list of options and figured out it belonged to the third bullet point. At least now it makes sense to me. If I'm wrong, please think about rewriting this section because it confused me. If you explain it to me I'm more than happy to rephrase it!